+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Preserve `Eloquent\Builder` when using filterColumn api. Allows us to use model scope and any eloquent magics.
 - Fractal integration extracted to own plugin [laravel-datatables-fractal](https://github.com/yajra/laravel-datatables-fractal).
 - Raw output are always passed on transformer instance.
+- Object response is now the default output `public function make($mDataSupport = true)`.
 
 ### REMOVED
 - Remove `filterColumn` api magic query method in favor of closure.